use native crypto package instead of jwt
