--- examples/webhook/index.js
@@ -1,39 +1,0 @@
-require('dotenv').config();
-const { Defender } = require('@openzeppelin/defender-sdk');
-
-const SECRET = 'ece309aa-bb4c-4dd5-995d-ad35e83f9924';
-const SIGNATURE =
-  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJldmVudHMiOlt7Imhhc2giOiIweDgxNWU1NzJiNWU5ZDUzMDRjZTQ0YjFhYTZjN2YyOWQzNmU1MWY4ZDMwMzM5YmYzN2U4YmEwMWE4YzUzNjRjMTYiLCJ0cmFuc2FjdGlvbiI6eyJ0cmFuc2FjdGlvbkhhc2giOiIweDgxNWU1NzJiNWU5ZDUzMDRjZTQ0YjFhYTZjN2YyOWQzNmU1MWY4ZDMwMzM5YmYzN2U4YmEwMWE4YzUzNjRjMTYiLCJibG9ja0hhc2giOiIweGFmOTUzYmVhYThhOGMxNmZjZDVmOGRkNjg2NzEyYWE0MjNjMTgxNTc2ZTliYzEyNGZlMzQxYmM5ZmNhNjk3ZGQiLCJibG9ja051bWJlciI6IjB4NWU3NmFlIiwibG9nc0Jsb29tIjoiMHg0MDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDkwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAyMDAwMDAwMjAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDQwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMCIsImdhc1VzZWQiOiIweDZiNjUiLCJjb250cmFjdEFkZHJlc3MiOm51bGwsImN1bXVsYXRpdmVHYXNVc2VkIjoiMHg3NmVmMGYiLCJ0cmFuc2FjdGlvbkluZGV4IjoiMHgyZCIsImZyb20iOiIweGNiOTRkZmE3ODIwZGNiM2VjMWZlNmEwYTZkODNlMzY1MzY4MGQyZDEiLCJ0byI6IjB4MjNhNjUzM2UyOTJmODNhY2Q3YWQ4Y2QwMmQwM2I2ZTBmYWNiYWM1MyIsInR5cGUiOiIweDIiLCJlZmZlY3RpdmVHYXNQcmljZSI6IjB4NWZkN2E1ZDk5IiwibG9ncyI6W3siYmxvY2tIYXNoIjoiMHhhZjk1M2JlYWE4YThjMTZmY2Q1ZjhkZDY4NjcxMmFhNDIzYzE4MTU3NmU5YmMxMjRmZTM0MWJjOWZjYTY5N2RkIiwiYWRkcmVzcyI6IjB4MjNhNjUzM2UyOTJmODNhY2Q3YWQ4Y2QwMmQwM2I2ZTBmYWNiYWM1MyIsImxvZ0luZGV4IjoiMHg2ZiIsImRhdGEiOiIweDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMGEiLCJyZW1vdmVkIjpmYWxzZSwidG9waWNzIjpbIjB4OTNmZTZkMzk3Yzc0ZmRmMTQwMmE4YjcyZTQ3YjY4NTEyZjA1MTBkN2I5OGE0YmM0Y2JkZjZhYzcxMDhiM2M1OSJdLCJibG9ja051bWJlciI6IjB4NWU3NmFlIiwidHJhbnNhY3Rpb25JbmRleCI6IjB4MmQiLCJ0cmFuc2FjdGlvbkhhc2giOiIweDgxNWU1NzJiNWU5ZDUzMDRjZTQ0YjFhYTZjN2YyOWQzNmU1MWY4ZDMwMzM5YmYzN2U4YmEwMWE4YzUzNjRjMTYifV0sInN0YXR1cyI6IjB4MSJ9LCJibG9ja0hhc2giOiIweGFmOTUzYmVhYThhOGMxNmZjZDVmOGRkNjg2NzEyYWE0MjNjMTgxNTc2ZTliYzEyNGZlMzQxYmM5ZmNhNjk3ZGQiLCJibG9ja051bWJlciI6IjB4NWU3NmFlIiwidGltZXN0YW1wIjoxNzE5NDA1NzU2LCJtYXRjaFJlYXNvbnMiOlt7InR5cGUiOiJldmVudCIsInNpZ25hdHVyZSI6IlZhbHVlQ2hhbmdlZCh1aW50MjU2KSIsImFkZHJlc3MiOiIweDIzYTY1MzNlMjkyZjgzYWNkN2FkOGNkMDJkMDNiNmUwZmFjYmFjNTMiLCJhcmdzIjpbIjEwIl0sInBhcmFtcyI6eyJ2YWx1ZSI6IjEwIn19LHsidHlwZSI6ImZ1bmN0aW9uIiwic2lnbmF0dXJlIjoic3RvcmUodWludDI1NikiLCJhZGRyZXNzIjoiMHgyM2E2NTMzZTI5MmY4M2FjZDdhZDhjZDAyZDAzYjZlMGZhY2JhYzUzIiwiYXJncyI6WyIxMCJdLCJwYXJhbXMiOnsidmFsdWUiOiIxMCJ9fV0sIm1hdGNoZWRBZGRyZXNzZXMiOlsiMHgyM2E2NTMzZTI5MmY4M2FjZDdhZDhjZDAyZDAzYjZlMGZhY2JhYzUzIl0sIm1hdGNoZWRDaGVja3N1bUFkZHJlc3NlcyI6WyIweDIzYTY1MzNFMjkyZjgzQWNEN2FEOGNEMDJEMDNCNkUwZmFDYkFDNTMiXSwic2VudGluZWwiOnsiaWQiOiI0YTliYWU3NC1jMGVkLTRjM2UtYmMwMC00NjUzZTllNDI5NzIiLCJuYW1lIjoiVGVzdCIsImFiaSI6W3sidHlwZSI6ImV2ZW50IiwiYW5vbnltb3VzIjpmYWxzZSwibmFtZSI6IlZhbHVlQ2hhbmdlZCIsImlucHV0cyI6W3sidHlwZSI6InVpbnQyNTYiLCJuYW1lIjoidmFsdWUiLCJpbmRleGVkIjpmYWxzZX1dfSx7InR5cGUiOiJmdW5jdGlvbiIsIm5hbWUiOiJyZXRyaWV2ZSIsImNvbnN0YW50Ijp0cnVlLCJzdGF0ZU11dGFiaWxpdHkiOiJ2aWV3IiwicGF5YWJsZSI6ZmFsc2UsImlucHV0cyI6W10sIm91dHB1dHMiOlt7InR5cGUiOiJ1aW50MjU2In1dfSx7InR5cGUiOiJmdW5jdGlvbiIsIm5hbWUiOiJzdG9yZSIsImNvbnN0YW50IjpmYWxzZSwicGF5YWJsZSI6ZmFsc2UsImlucHV0cyI6W3sidHlwZSI6InVpbnQyNTYiLCJuYW1lIjoidmFsdWUifV0sIm91dHB1dHMiOltdfV0sImFkZHJlc3NlcyI6WyIweDIzYTY1MzNFMjkyZjgzQWNEN2FEOGNEMDJEMDNCNkUwZmFDYkFDNTMiXSwiY29uZmlybUJsb2NrcyI6MSwibmV0d29yayI6InNlcG9saWEiLCJjaGFpbklkIjoxMTE1NTExMX0sIm1vbml0b3IiOnsiaWQiOiI0YTliYWU3NC1jMGVkLTRjM2UtYmMwMC00NjUzZTllNDI5NzIiLCJuYW1lIjoiVGVzdCIsImFiaSI6W3sidHlwZSI6ImV2ZW50IiwiYW5vbnltb3VzIjpmYWxzZSwibmFtZSI6IlZhbHVlQ2hhbmdlZCIsImlucHV0cyI6W3sidHlwZSI6InVpbnQyNTYiLCJuYW1lIjoidmFsdWUiLCJpbmRleGVkIjpmYWxzZX1dfSx7InR5cGUiOiJmdW5jdGlvbiIsIm5hbWUiOiJyZXRyaWV2ZSIsImNvbnN0YW50Ijp0cnVlLCJzdGF0ZU11dGFiaWxpdHkiOiJ2aWV3IiwicGF5YWJsZSI6ZmFsc2UsImlucHV0cyI6W10sIm91dHB1dHMiOlt7InR5cGUiOiJ1aW50MjU2In1dfSx7InR5cGUiOiJmdW5jdGlvbiIsIm5hbWUiOiJzdG9yZSIsImNvbnN0YW50IjpmYWxzZSwicGF5YWJsZSI6ZmFsc2UsImlucHV0cyI6W3sidHlwZSI6InVpbnQyNTYiLCJuYW1lIjoidmFsdWUifV0sIm91dHB1dHMiOltdfV0sImFkZHJlc3NlcyI6WyIweDIzYTY1MzNFMjkyZjgzQWNEN2FEOGNEMDJEMDNCNkUwZmFDYkFDNTMiXSwiY29uZmlybUJsb2NrcyI6MSwibmV0d29yayI6InNlcG9saWEiLCJjaGFpbklkIjoxMTE1NTExMX0sInR5cGUiOiJCTE9DSyIsInZhbHVlIjoiMHgwIn1dLCJpYXQiOjE3MTk0MDY5NDV9.DUPgOv6D3L_r34BfEQ_8nFhCq_5Xp9V0WzWCt9OL4Rw';
-
-function webhook(req, res) {
-  const creds = {
-    apiKey: process.env.API_KEY,
-    apiSecret: process.env.API_SECRET,
-  };
-  const client = new Defender(creds);
-
-  const valid = client.notificationChannel.verifySignature({
-    secret: SECRET,
-    signature: req.signature,
-  });
-
-  res.json({ message: valid ? 'valid signature' : 'invalid signature' });
-}
-
-function main() {
-  // example webhook payload
-  const req = {
-    events: [],
-    signature: SIGNATURE,
-  };
-  const res = { json: console.log };
-  webhook(req, res);
-}
-
-if (require.main === module) {
-  try {
-    main();
-  } catch (error) {
-    console.error(error);
-  }
-}
